+++ packages/sdk/src/client/index.ts
@@ -26,10 +26,14 @@ export type {
 	S3ProviderDefinition,
 	HasRequiredInput,
 	WithInput,
+	ExtractProfileName,
+	ExtractMeta,
 } from './types';
 
+export type { UploadValidationOptions } from './client';
 export { Client } from './client';
 
-export { GraphQLResponseError } from './GraphQLResponseError';
-export { ResponseError } from './ResponseError';
 export type { ClientResponseError } from './ClientResponseError';
+export { GraphQLResponseError } from './GraphQLResponseError';
+export { InputValidationError } from './InputValidationError';
+export { ResponseError } from './ResponseError';