+++ packages/sdk/src/codegen/templates/typescript/internal.client.template.ts
@@ -8,8 +8,6 @@ export interface Queries  {
     {{#each internalQueries}}
         {{operationName}}: ({{#if hasInternalInput}}options: OperationArgsWithInput<Internal{{operationName}}Input>{{/if}}) => Promise<{{operationName}}Response>;
     {{/each}}
-{{else}}
-  queries: {};
 {{/if}}
 }
 
@@ -18,8 +16,6 @@ export interface Mutations  {
     {{#each internalMutations}}
         {{operationName}}: ({{#if hasInternalInput}}options: OperationArgsWithInput<{{operationName}}Input>{{/if}}) => Promise<{{operationName}}Response>;
     {{/each}}
-{{else}}
-  mutations: {};
 {{/if}}
 }