+++ packages/sdk/src/v2openapi/index.ts
@@ -49,7 +49,7 @@ import { EnvironmentVariable, InputVariable, mapInputVariable } from '../configu
 import { HeadersBuilder, mapHeaders } from '../definition/headers-builder';
 import { Logger } from '../logger';
 import _ from 'lodash';
-import transformSchema from '../transformations/schema';
+import transformSchema from '../transformations/transformSchema';
 
 export const openApiSpecificationToRESTApiObject = async (
 	oas: string,
@@ -338,6 +338,7 @@ class RESTApiBuilder {
 							renderConfiguration: ArgumentRenderConfiguration.RENDER_ARGUMENT_AS_ARRAY_CSV,
 							sourceType: ArgumentSource.FIELD_ARGUMENT,
 							sourcePath: [param.name],
+							renameTypeTo: '',
 						});
 					}
 					break;