+++ pkg/engineconfigloader/engineconfigloader.go
@@ -264,7 +264,7 @@ func New(wundergraphDir string, resolvers ...FactoryResolver) *EngineConfigLoade
 	}
 }
 
-func (l *EngineConfigLoader) Load(engineConfig wgpb.EngineConfiguration, hooksServerUrl string) (*plan.Configuration, error) {
+func (l *EngineConfigLoader) Load(engineConfig wgpb.EngineConfiguration) (*plan.Configuration, error) {
 	var (
 		outConfig plan.Configuration
 	)
@@ -274,8 +274,9 @@ func (l *EngineConfigLoader) Load(engineConfig wgpb.EngineConfiguration, hooksSe
 		var args []plan.ArgumentConfiguration
 		for _, argumentConfiguration := range configuration.ArgumentsConfiguration {
 			arg := plan.ArgumentConfiguration{
-				Name:       argumentConfiguration.Name,
-				SourcePath: argumentConfiguration.SourcePath,
+				Name:         argumentConfiguration.Name,
+				SourcePath:   argumentConfiguration.SourcePath,
+				RenameTypeTo: argumentConfiguration.RenameTypeTo,
 			}
 			switch argumentConfiguration.SourceType {
 			case wgpb.ArgumentSource_FIELD_ARGUMENT:
@@ -347,7 +348,7 @@ func (l *EngineConfigLoader) Load(engineConfig wgpb.EngineConfiguration, hooksSe
 
 			fetchURL := buildFetchUrl(
 				loadvariable.String(in.CustomRest.Fetch.GetUrl()),
-				baseUrl(loadvariable.String(in.CustomRest.Fetch.GetBaseUrl()), hooksServerUrl),
+				loadvariable.String(in.CustomRest.Fetch.GetBaseUrl()),
 				loadvariable.String(in.CustomRest.Fetch.GetPath()))
 
 			// resolves arguments like {{ .arguments.tld }} are allowed
@@ -391,7 +392,7 @@ func (l *EngineConfigLoader) Load(engineConfig wgpb.EngineConfiguration, hooksSe
 
 			fetchUrl := buildFetchUrl(
 				loadvariable.String(in.CustomGraphql.Fetch.GetUrl()),
-				baseUrl(loadvariable.String(in.CustomGraphql.Fetch.GetBaseUrl()), hooksServerUrl),
+				loadvariable.String(in.CustomGraphql.Fetch.GetBaseUrl()),
 				loadvariable.String(in.CustomGraphql.Fetch.GetPath()),
 			)
 
@@ -530,13 +531,4 @@ func buildFetchUrl(url, baseUrl, path string) string {
 	}
 
 	return fmt.Sprintf("%s/%s", strings.TrimSuffix(baseUrl, "/"), strings.TrimPrefix(path, "/"))
-}
-
-const serverUrlPlaceholder = "WG_SERVER_URL"
-
-func baseUrl(baseUrl, serverUrl string) string {
-	if baseUrl == serverUrlPlaceholder {
-		return serverUrl
-	}
-	return baseUrl
 }