+++ testapps/default/.wundergraph/wundergraph.config.ts
@@ -30,23 +30,23 @@ const jsp2 = introspect.openApi({
 	headers: (builder) => builder.addClientRequestHeader('X-Authorization', 'Authorization'),
 });
 
-/*const federatedApi = introspect.federation({
+const federatedApi = introspect.federation({
 	apiNamespace: 'federated',
 	upstreams: [
 		{
-			url: 'http://localhost:4001/graphql',
+			url: 'https://wg-federation-demo-accounts.fly.dev/graphql',
 		},
 		{
-			url: 'http://localhost:4002/graphql',
+			url: 'https://wg-federation-demo-products.fly.dev/graphql',
 		},
 		{
-			url: 'http://localhost:4003/graphql',
+			url: 'https://wg-federation-demo-reviews.fly.dev/graphql',
 		},
 		{
-			url: 'http://localhost:4004/graphql',
+			url: 'https://wg-federation-demo-inventory.fly.dev/graphql',
 		},
 	],
-});*/
+});
 
 const spacex = introspect.graphql({
 	apiNamespace: 'spacex',
@@ -86,7 +86,7 @@ const usersPost = introspect.prisma({
 
 // configureWunderGraph emits the configuration
 configureWunderGraphApplication({
-	apis: [jsp, weather, countries, spacex, chinook, db, jsp2, usersPost],
+	apis: [jsp, weather, countries, spacex, chinook, db, jsp2, usersPost, federatedApi],
 	server,
 	operations,
 	authorization: {
@@ -122,9 +122,6 @@ configureWunderGraphApplication({
 		cookieBased: {
 			providers: [authProviders.demo()],
 			authorizedRedirectUris: ['http://localhost:3000'],
-			secureCookieHashKey: new EnvironmentVariable('WUNDERGRAPH_SECURE_COOKIE_HASH_KEY'), // must be of length 32
-			secureCookieBlockKey: new EnvironmentVariable('WUNDERGRAPH_SECURE_COOKIE_BLOCK_KEY'), // must be of length 32
-			csrfTokenSecret: new EnvironmentVariable('WUNDERGRAPH_CSRF_TOKEN_SECRET'), // must be of length 11
 		},
 	},
 	security: {