+++ testapps/default/jest.config.js
@@ -4,5 +4,6 @@ module.exports = {
 		'^.+\\.tsx?$': 'ts-jest',
 	},
 	testRegex: '(/__tests__/.*|\\.(test|spec))\\.[tj]sx?$',
+	testTimeout: 10 * 1000,
 	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
 };